Disable APA 10.2:32
Stops publisher from being suppressed in @Book when its value
contains the author.

It is tricky to override \DeclareStyleSourcemap from apa.bbx, so
patch it in the .deb.

Note that although the comment in apa.bbx says "Remove PUBLISHER if it
is the same as AUTHOR", publisher is actually removed if it contains
author.